backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart
python-dotenv
httpx>=0.25.2

# Database
aiosqlite

# Computer Vision (Gemini-only — no YOLO)
opencv-python>=4.9.0
numpy>=2.0.0
Pillow>=10.3.0

# Traffic Analytics (truck flow monitoring)
ultralytics>=8.3.0

# LLM
openai

# Optional
# open-clip-torch
